storageTask/requirements.txt
gevent==1.0.2
PyYAML==3.11